chore: single modal for strategies and release plan preview
