+++ frontend/src/component/feature/FeatureView/FeatureOverview/FeatureOverviewEnvironments/FeatureOverviewEnvironment/FeatureOverviewEnvironment.tsx
@@ -109,7 +109,6 @@ export const FeatureOverviewEnvironment = ({
                             featureId={featureId}
                             environmentId={environment.name}
                             variant='outlined'
-                            size='small'
                         />
                     ) : (
                         <FeatureOverviewEnvironmentMetrics
